move malformed pipeline test
